requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
pandas>=1.5.0
tabulate>=0.9.0
